requirements.txt
django-allauth[mfa,saml,socialaccount]==0.60.1
djangorestframework